[#34643] Separate Show & Edit Forms
